move named tuple querable to new file
