+++ scalasql/namedtuples/test/src/example/SimpleTableH2Example.scala
@@ -2,7 +2,7 @@
 package scalasql.namedtuples.example
 
 import scalasql.namedtuples.SimpleTable
-import scalasql.namedtuples.SimpleTable.NamedTupleOps.given
+import scalasql.namedtuples.NamedTupleQueryable.given
 import scalasql.H2Dialect.*
 
 object SimpleTableH2Example {
