+++ scalasql/namedtuples/test/src/example/SimpleTablePostgresExample.scala
@@ -3,7 +3,7 @@ package scalasql.namedtuples.example
 
 import org.testcontainers.containers.PostgreSQLContainer
 import scalasql.namedtuples.SimpleTable
-import scalasql.namedtuples.SimpleTable.NamedTupleOps.given
+import scalasql.namedtuples.NamedTupleQueryable.given
 
 import scalasql.PostgresDialect.*
